Finish styling "results" section and start styling "summary" section.
Finished the styles on the "results" section which included keeping
styles static when the page is zoomed in. This required adding
"min-height" properties to .component and .result - allowing them to
grow to changing screen sizes and zooms. Stopping to learn more about
CSS grid before moving onto the styles for .summary.